rename: create a Footer folder and move the footer files
